Updates
Added option in the Settings file for a different ntuple name in the
input root file. The default is still what Geant4 outputs. Also methods
that check that all the "crystal" hits are unique, that is, they have
different crystal and detector IDs. If they have the same crystal and
detector IDs, then we sum the energies together. Normally the Geant4
simulation would sum energy deposits on the same volume, but if we ran
the code in "step mode", or if we merged two ntuples together, this
would not be true. These methods check and will do what "hit mode" in
Geant4 normally does for us!