Removing the DescantNeutronDiscrimination "hack"
The DescantNeutronDiscrimination method screws up the switch statement.
In the case the particle was not a neutron the if statement is not
satisfied, and descends into the next case. This
DescantNeutronDiscrimination method, was a temporary hack anyway, it
reality we would like to look at the DESCANT spectrum to determine the
particle type.